Refactor SQLite library loading and database initialization
This commit refactors the code related to loading the appropriate SQLite library based on the platform and initializing the database. The code for loading the library and opening the database has been moved to a separate file called `connection.dart`. The `loadSqlite3Flutter` and `_loadSecureStorageLibrary` methods have been removed from `DatabaseProvider` and replaced with calls to the new methods in `connection.dart`. The code for opening the SQLite library on iOS, Android, and Linux has also been moved to the respective platform-specific methods in `connection.dart`. Additionally, the `openRecipeDatabase` method in `unsupported.dart` now throws an `UnsupportedError` with a message indicating that no database implementation was found.